+++ scripts/nvt-scan.gmp.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2017-2019 Greenbone Networks GmbH
+# Copyright (C) 2017-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,6 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import sys
 from datetime import datetime
 
 from gvm.errors import GvmError
@@ -36,7 +37,7 @@ def check_args(args):
     ssh --hostname <gsm> 1.3.6.1.4.1.25623.1.0.106223 localhost
         """
         print(message)
-        quit()
+        sys.exit()
 
 
 def create_config(gmp, nvt_oid):