+++ CHANGELOG.md
@@ -25,6 +25,8 @@ $ cd gvm-tools && git log
 * CI tests Python 3.9 now. [#353](https://github.com/greenbone/gvm-tools/pull/353)
 
 ### Changed
+* Moved generic functions to generate random ids and ips from scripts to the `helper` module. [#365](https://github.com/greenbone/gvm-tools/pull/365)
+
 ### Deprecated
 * Dropped Python 3.5 and Python 3.6 support. Python 3.7+ is required now. [#353](https://github.com/greenbone/gvm-tools/pull/353)